add folder contacts to controllers: - by dividing separately function
