--- controllers/contacts.js
@@ -1,58 +1,0 @@
-const { Contact } = require("../models/contact");
-
-const { HttpError, ctrlWrapper } = require("../helpers");
-
-const get = async (req, res) => {
-  const result = await Contact.find({}, "-createdAt -updatedAt");
-  res.json(result);
-};
-
-const getById = async (req, res) => {
-  const { id } = req.params;
-  const result = await Contact.findById(id);
-  if (!result) {
-    throw HttpError(404, "Not found!");
-  }
-  res.json(result);
-};
-
-const add = async (req, res) => {
-  const result = await Contact.create(req.body);
-  res.status(201).json(result);
-};
-
-const deleteById = async (req, res) => {
-  const { id } = req.params;
-  const result = await Contact.findByIdAndRemove(id);
-  if (!result) {
-    throw HttpError(404, "Not Found!");
-  }
-  res.status(200).json({ message: "contact deleted" });
-};
-
-const updateById = async (req, res) => {
-  const { id } = req.params;
-  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
-  if (!result) {
-    throw HttpError(404, "Not Found!");
-  }
-  res.status(200).json(result);
-};
-
-const updateFavorite = async (req, res) => {
-  const { id } = req.params;
-  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
-  if (!result) {
-    throw HttpError(404, "Not Found!");
-  }
-  res.status(200).json(result);
-};
-
-module.exports = {
-  get: ctrlWrapper(get),
-  getById: ctrlWrapper(getById),
-  add: ctrlWrapper(add),
-  updateById: ctrlWrapper(updateById),
-  updateFavorite: ctrlWrapper(updateFavorite),
-  deleteById: ctrlWrapper(deleteById),
-};
