+++ controllers/contacts/add.js
@@ -1,0 +1,8 @@
+const { Contact } = require("../../models/contact");
+
+const add = async (req, res) => {
+  const result = await Contact.create(req.body);
+  res.status(201).json(result);
+};
+
+module.exports = add;
