+++ controllers/contacts/deleteById.js
@@ -1,0 +1,12 @@
+const { Contact } = require("../../models/contact");
+
+const deleteById = async (req, res) => {
+  const { id } = req.params;
+  const result = await Contact.findByIdAndRemove(id);
+  if (!result) {
+    throw HttpError(404, "Not Found!");
+  }
+  res.status(200).json({ message: "contact deleted" });
+};
+
+module.exports = deleteById;
