+++ controllers/contacts/index.js
@@ -1,0 +1,17 @@
+const get = require("./get");
+const getById = require("./getById");
+const add = require("./add");
+const deleteById = require("./deleteById");
+const updateById = require("./updateById");
+const updateFavorite = require("./updateFavorite");
+
+const { HttpError, ctrlWrapper } = require("../../helpers");
+
+module.exports = {
+  get: ctrlWrapper(get),
+  getById: ctrlWrapper(getById),
+  add: ctrlWrapper(add),
+  deleteById: ctrlWrapper(deleteById),
+  updateById: ctrlWrapper(updateById),
+  updateFavorite: ctrlWrapper(updateFavorite),
+};
