+++ controllers/contacts/updateFavorite.js
@@ -1,0 +1,12 @@
+const { Contact } = require("../../models/contact");
+
+const updateFavorite = async (req, res) => {
+  const { id } = req.params;
+  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  if (!result) {
+    throw HttpError(404, "Not Found!");
+  }
+  res.status(200).json(result);
+};
+
+module.exports = updateFavorite;
